requirements.txt
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
pur==7.3.3
pika==1.3.2
